[#82] Refactor: Pull up Field; pulled up receiverUid and timeRead to base Notification as they are relevant variable for any Notification
