+++ app/src/main/java/com/example/bigowlapp/model/ScheduleRequest.java
@@ -1,14 +1,11 @@
 package com.example.bigowlapp.model;
 
-import com.google.firebase.Timestamp;
 import com.google.firebase.firestore.IgnoreExtraProperties;
 
 @IgnoreExtraProperties
 public class ScheduleRequest extends Notification {
     private String senderUid;
-    private String receiverUid;
     private String groupUid;
-    private Timestamp timeRead;
     private UserScheduleResponse senderResponse;
 
     public ScheduleRequest() {
@@ -23,28 +20,12 @@ public class ScheduleRequest extends Notification {
         this.senderUid = senderUid;
     }
 
-    public String getReceiverUid() {
-        return receiverUid;
-    }
-
-    public void setReceiverUid(String receiverUid) {
-        this.receiverUid = receiverUid;
-    }
-
     public String getGroupUid() {
         return groupUid;
     }
 
     public void setGroupUid(String groupUid) {
         this.groupUid = groupUid;
-    }
-
-    public Timestamp getTimeRead() {
-        return timeRead;
-    }
-
-    public void setTimeRead(Timestamp timeRead) {
-        this.timeRead = timeRead;
     }
 
     public UserScheduleResponse getSenderResponse() {
@@ -55,13 +36,9 @@ public class ScheduleRequest extends Notification {
         this.senderResponse = senderResponse;
     }
 
-    public static class Field extends Notification.Field{
+    public static class Field extends Notification.Field {
         public static final String SENDER_UID = "senderUid";
-        // TODO: should probably pull up this field as its relevant to all Notifications
-        public static final String RECEIVER_UID = "receiverUid";
         public static final String GROUP_UID = "groupUid";
-        // TODO: should probably pull up this field as its relevant to all Notifications
-        public static final String TIME_READ = "timeRead";
         public static final String SENDER_RESPONSE = "senderResponse";
     }
 }
