+++ app/src/main/java/com/example/bigowlapp/model/SupervisionRequest.java
@@ -62,9 +62,8 @@ public class SupervisionRequest extends Notification {
         NEUTRAL
     }
 
-    public static class Field extends Notification.Field{
+    public static class Field extends Notification.Field {
         public static final String SENDER_UID = "senderUid";
-        public static final String RECEIVER_UID = "receiverUid";
         public static final String GROUP_UID = "groupUid";
         public static final String RESPONSE = "response";
         public static final String TIME_RESPONSE = "timeResponse";
